spilt mesh provider implementation.
