+++ pkg/constants/constants-ossm.go
@@ -1,0 +1,7 @@
+package constants
+
+// openshift service mesh constants
+const (
+	FederationEgressServiceLabelKey  = "federation.maistra.io/egress-for"
+	FederationIngressServiceLabelKey = "federation.maistra.io/ingress-for"
+)
