+++ pkg/constants/constants.go
@@ -9,20 +9,6 @@ const (
 	AnnotationKeyForMeshFederationOwner = "mesh.open-cluster-management.io/federation-owner"
 )
 
-// common istio constants
-const (
-	IstioCAConfigmapName  = "istio-ca-root-cert"
-	IstioCAConfigmapKey   = "root-cert.pem"
-	IstioCAConfigmapLabel = "istio.io/config"
-	IstioCASecretName     = "cacerts"
-)
-
-// openshift service mesh constants
-const (
-	FederationEgressServiceLabelKey  = "federation.maistra.io/egress-for"
-	FederationIngressServiceLabelKey = "federation.maistra.io/ingress-for"
-)
-
 // mesh federation configuration constants
 const (
 	FederationConfigMapMeshPeerCALabelKey          = "root-cert.pem"
