+++ pkg/utils/utils.go
@@ -2,6 +2,7 @@ package utils
 
 import "k8s.io/client-go/kubernetes"
 
+// SliceContainsString check if the string slice contains the given string
 func SliceContainsString(s []string, str string) bool {
 	for _, v := range s {
 		if v == str {
@@ -11,6 +12,7 @@ func SliceContainsString(s []string, str string) bool {
 	return false
 }
 
+// IsOpenshift check if the cluster is openshift cluster with given kubeclient
 func IsOpenshift(kubeclient kubernetes.Interface) bool {
 	serverGroups, err := kubeclient.Discovery().ServerGroups()
 	if err != nil {
